+++ pylinkage/interface/__init__.py
@@ -3,7 +3,8 @@ Classes that represent the main interface with the library.
 """
 from .exceptions import (
     UnbuildableError,
-    HypostaticError
+    HypostaticError,
+    NotCompletelyDefinedError,
 )
 from .joint import (
     Static,