+++ tests/test_geometry.py
@@ -7,8 +7,9 @@ Created on Thu Apr 15, 13:26:42 2021
 """
 
 import unittest
-from pylinkage.geometry import circle_intersect, circle_line_intersection
 import numpy as np
+
+from pylinkage.geometry import circle_intersect, circle_line_intersection, sqr_dist
 
 
 class TestCircles(unittest.TestCase):
@@ -67,7 +68,7 @@ def circle_line_intersection_data(mode):
             line[:2],
             vector * circle[2] * np.random.rand()
         )
-    return circle, tuple(line)
+    return circle, tuple(map(float, line))
 
 
 class TestCircleLineIntersection(unittest.TestCase):
@@ -90,8 +91,10 @@ class TestCircleLineIntersection(unittest.TestCase):
         count = 0
         for _ in range(20):
             circle, line = circle_line_intersection_data(1)
-            intersection = circle_line_intersection(circle, tuple(line))
+            intersection = circle_line_intersection(circle, line)
             if len(intersection) == 1:
+                count += 1
+            elif len(intersection) == 2 and 1e-5 > sqr_dist(intersection[0], intersection[1]):
                 count += 1
         self.assertLessEqual(
             7, count, f"Success rate of {count * 5}% is too low!"